For both boost::uint64_t & std::uint64_t , the typedef is unsigned long long. Using it explicitly.

+++ modules/tinydnsbackend/tinydnsbackend.cc
@@ -260,11 +260,11 @@ bool TinyDNSBackend::get(DNSResourceRecord &rr)
 			rr.auth = true;
 
 			rr.ttl = pr.get32BitInt();
-			long timestamp = pr.get32BitInt();
+			unsigned long long timestamp = pr.get32BitInt();
 			timestamp <<= 32;
 			timestamp += pr.get32BitInt();
 			if(timestamp) {
-				long now = d_taiepoch + time(NULL);
+				unsigned long long now = d_taiepoch + time(NULL);
 				if (rr.ttl == 0) {
 					if (timestamp < now) {
 						continue;
